extract module moved, fixed lints and formatting
